+++ packages/date-picker/src/FilterDateRangePicker/components/Trigger/index.ts
@@ -1,1 +1,2 @@
 export * from "./FilterTriggerButton"
+export * from "./RemovableFilterTriggerButton"