+++ packages/component-library/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [16.1.11](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/component-library@16.1.10...@kaizen/component-library@16.1.11) (2022-12-01)
+
+**Note:** Version bump only for package @kaizen/component-library
+
+
+
+
+
 ## [16.1.10](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/component-library@16.1.9...@kaizen/component-library@16.1.10) (2022-11-25)
 
 **Note:** Version bump only for package @kaizen/component-library